Machine Learning Model
In this repository, a model is built using machine learning algorithm where based on the details of a person has been filled, one can come to know the chances of liver disease, like positive chances where as negative chances. In this model algorithm  selected was Random forest based on the maximum accuracy gained among all algorithms.